CodeEditor component merged into single file
